+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 2.90.0
+
+This release adds support for wrapping PEP-660 `build_editable` to `pex_build.setuptools.build`
+plugins and dogfoods this.
+
+* Support wrapping `build_editable` in wrap. (#3105)
+
 ## 2.89.1
 
 This release adds better diagnostics for certain Pex filesystem interaction errors.